+++ crates/fuel-core/src/service/adapters/fuel_gas_price_provider.rs
@@ -4,10 +4,13 @@ use fuel_core_gas_price_service::{
     SharedGasPriceAlgo,
 };
 use fuel_core_producer::block_producer::gas_price::GasPriceProvider as ProducerGasPriceProvider;
-use fuel_core_txpool::ports::GasPriceProvider as TxPoolGasPricProvider;
+use fuel_core_txpool::ports::GasPriceProvider as TxPoolGasPriceProvider;
 use fuel_core_types::{
     fuel_types::BlockHeight,
-    services::txpool::Result as TxPoolResult,
+    services::txpool::{
+        Error as TxPoolError,
+        Result as TxPoolResult,
+    },
 };
 
 pub type Result<T, E = Error> = std::result::Result<T, E>;
@@ -53,12 +56,8 @@ impl<A> FuelGasPriceProvider<A>
 where
     A: GasPriceAlgorithm + Send + Sync,
 {
-    async fn next_gas_price(&self, block_bytes: u64) -> u64 {
-        self.algorithm.next_gas_price(block_bytes).await
-    }
-
-    async fn last_gas_price(&self) -> u64 {
-        self.algorithm.last_gas_price().await
+    async fn next_gas_price(&self) -> Option<u64> {
+        self.algorithm.next_gas_price().await
     }
 }
 
@@ -67,18 +66,24 @@ impl<A> ProducerGasPriceProvider for FuelGasPriceProvider<A>
 where
     A: GasPriceAlgorithm + Send + Sync,
 {
-    async fn next_gas_price(&self, block_bytes: u64) -> anyhow::Result<u64> {
-        Ok(self.next_gas_price(block_bytes).await)
+    async fn next_gas_price(&self) -> anyhow::Result<u64> {
+        self.next_gas_price()
+            .await
+            .ok_or(anyhow::anyhow!("No gas price available"))
     }
 }
 
 #[async_trait::async_trait]
-impl<A> TxPoolGasPricProvider for FuelGasPriceProvider<A>
+impl<A> TxPoolGasPriceProvider for FuelGasPriceProvider<A>
 where
     A: GasPriceAlgorithm + Send + Sync,
 {
-    async fn last_gas_price(&self) -> TxPoolResult<u64> {
-        Ok(self.last_gas_price().await)
+    async fn next_gas_price(&self) -> TxPoolResult<u64> {
+        self.next_gas_price()
+            .await
+            .ok_or(TxPoolError::GasPriceNotFound(
+                "Gas price not set yet".to_string(),
+            ))
     }
 }
 
@@ -87,7 +92,7 @@ impl<A> GraphqlGasPriceEstimate for FuelGasPriceProvider<A>
 where
     A: GasPriceAlgorithm + Send + Sync,
 {
-    async fn worst_case_gas_price(&self, height: BlockHeight) -> u64 {
+    async fn worst_case_gas_price(&self, height: BlockHeight) -> Option<u64> {
         self.algorithm.worst_case_gas_price(height).await
     }
 }