+++ crates/services/gas_price_service/src/fuel_gas_price_updater/fuel_core_storage_adapter.rs
@@ -87,8 +87,23 @@ where
 }
 
 pub struct FuelL2BlockSource<Settings> {
+    genesis_block_height: BlockHeight,
     gas_price_settings: Settings,
     committed_block_stream: BoxStream<SharedImportResult>,
+}
+
+impl<Settings> FuelL2BlockSource<Settings> {
+    pub fn new(
+        genesis_block_height: BlockHeight,
+        gas_price_settings: Settings,
+        committed_block_stream: BoxStream<SharedImportResult>,
+    ) -> Self {
+        Self {
+            genesis_block_height,
+            gas_price_settings,
+            committed_block_stream,
+        }
+    }
 }
 
 #[derive(Debug, Clone, PartialEq)]
@@ -110,9 +125,9 @@ fn get_block_info(
 ) -> GasPriceResult<BlockInfo> {
     let (fee, gas_price) = mint_values(block)?;
     let height = *block.header().height();
-    let calculated_used_gas = block_used_gas(height, fee, gas_price, gas_price_factor)?;
-    let used_gas = min(calculated_used_gas, block_gas_limit);
-    let info = BlockInfo {
+    let used_gas =
+        block_used_gas(height, fee, gas_price, gas_price_factor, block_gas_limit)?;
+    let info = BlockInfo::Block {
         height: (*block.header().height()).into(),
         gas_used: used_gas,
         block_gas_capacity: block_gas_limit,
@@ -136,6 +151,7 @@ fn block_used_gas(
     fee: u64,
     gas_price: u64,
     gas_price_factor: u64,
+    max_used_gas: u64,
 ) -> GasPriceResult<u64> {
     let scaled_fee =
         fee.checked_mul(gas_price_factor)
@@ -145,12 +161,10 @@ fn block_used_gas(
                     "Failed to scale fee by gas price factor, overflow"
                 ),
             })?;
-    scaled_fee
-        .checked_div(gas_price)
-        .ok_or(GasPriceError::CouldNotFetchL2Block {
-            block_height,
-            source_error: anyhow!("Failed to calculate gas used, division by zero"),
-        })
+    // If gas price is zero, assume max used gas
+    let approximate = scaled_fee.checked_div(gas_price).unwrap_or(max_used_gas);
+    let used_gas = min(approximate, max_used_gas);
+    Ok(used_gas)
 }
 
 #[async_trait::async_trait]
@@ -172,12 +186,21 @@ where
             .sealed_block
             .entity;
 
-        let param_version = block.header().consensus_parameters_version;
+        match block.header().height().cmp(&self.genesis_block_height) {
+            std::cmp::Ordering::Less => Err(GasPriceError::CouldNotFetchL2Block {
+                block_height: height,
+                source_error: anyhow!("Block precedes expected genesis block height"),
+            }),
+            std::cmp::Ordering::Equal => Ok(BlockInfo::GenesisBlock),
+            std::cmp::Ordering::Greater => {
+                let param_version = block.header().consensus_parameters_version;
 
-        let GasPriceSettings {
-            gas_price_factor,
-            block_gas_limit,
-        } = self.gas_price_settings.settings(&param_version)?;
-        get_block_info(block, gas_price_factor, block_gas_limit)
+                let GasPriceSettings {
+                    gas_price_factor,
+                    block_gas_limit,
+                } = self.gas_price_settings.settings(&param_version)?;
+                get_block_info(block, gas_price_factor, block_gas_limit)
+            }
+        }
     }
 }