+++ crates/services/producer/src/block_producer/tests.rs
@@ -63,7 +63,7 @@ impl MockProducerGasPrice {
 
 #[async_trait::async_trait]
 impl GasPriceProvider for MockProducerGasPrice {
-    async fn next_gas_price(&self, _block_bytes: u64) -> anyhow::Result<u64> {
+    async fn next_gas_price(&self) -> anyhow::Result<u64> {
         self.gas_price
             .ok_or_else(|| anyhow::anyhow!("Gas price not provided"))
     }